Update image, quantity to double

+++ BusinessEntity/OrderDetail.cs
@@ -20,7 +20,7 @@ namespace BusinessEntity
 
         public int CategoryId { get; set; }
 
-        public int Quanlity { get; set; }
+        public double Quanlity { get; set; }
 
         public string UnitName { get; set; }
